+++ src/cashbook/cashbook.console/Head.cs
@@ -47,7 +47,7 @@ namespace cashbook.console
 		[Verb]
 		void Withdraw (
 			[Required, Aliases("d")] 								DateTime transactionDate, 
-			[Required, Aliases("a")] 								double amount, 
+			[Required, Aliases("a")] 								decimal amount, 
 			[Required, Aliases("desc"), DefaultValue("Deposit")] 	string description,
 			[Aliases("f")] 											bool force) {
 			body.Withdraw (transactionDate, amount, description, force,
@@ -64,7 +64,7 @@ namespace cashbook.console
 		[Verb]
 		void Deposit (
 			[Required, Aliases("d")] 					DateTime transactionDate, 
-			[Required, Aliases("a")] 					double amount, 
+			[Required, Aliases("a")] 					decimal amount, 
 			[Aliases("desc"), DefaultValue("Deposit")] 	string description,
 			[Aliases("f")] 								bool force) {
 			body.Deposit (transactionDate, amount, description, force,
